refactor: replace ViewDatasetItem with DatasetItemFields for rendering dataset item details
